pyproject.toml
[project]
name = "echo"
version = "0.1.3"
description = "Framework-agnostic medical Agent building SDK with multi-provider LLM and Tools support"
readme = "README.md"
license = "MIT"
authors = [
    { name = "dicodetomy-paradox", email = "neha@eka.care" }
]
requires-python = ">=3.11"
keywords = ["agents", "llm", "tools", "mcp", "framework-agnostic", "multi-provider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.12.3",
    "orjson>=3.11.4",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/eka-care/echo-sdk"
Repository = "https://github.com/eka-care/echo-sdk"
Issues = "https://github.com/eka-care/echo-sdk/issues"

[project.optional-dependencies]
# LLM Providers
bedrock = ["boto3>=1.38.23"]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]

# Framework Integrations
crewai = ["crewai>=1.6.1"]
mcp = ["mcp>=1.8.0"]

# Convenience bundles
all = [
    "boto3>=1.38.23",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "crewai>=1.6.1",
    "mcp>=1.8.0",
]

dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
